Update to nhood_gr in _nhood_labels_to_cells_last_wins
In the old version, nhood_gr first dropped NA by all columns in the dataframe and then took unique values, without control of ordering by the user.
This change let´s the user control the ordering of groups, and removes the problem as described, by using pd.Categorical.
Reran tests, and it works.